[REF] prepare a big code refactor. First step moving all the code in the same file
